feat: extract shared component for profile edit page used by both client and admin
