+++ frontend/app/admin/(main)/myprofile/page.tsx
@@ -1,8 +1,2 @@
-export default function ProfilePage() {
-    return (
-      <div>
-        <h1>프로필</h1>
-        <p>프로필 페이지입니다!</p>
-      </div>
-    );
-  }
+import MyProfilePage from '@/components/MyProfilePage';
+export default MyProfilePage;
